% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneMeanDream-package.R
\docType{package}
\name{OneMeanDream-package}
\alias{OneMeanDream}
\alias{OneMeanDream-package}
\title{OneMeanDream: Enhanced One-Sample t-Test with Diagnostics}
\description{
Performs one-sample t-tests using Wallace critical value approximation and
Hastings p-value approximation. Includes comprehensive assumption checks
(normality via Shapiro-Wilk, outlier detection) and diagnostic plots.
}
\section{Main Functions}{

\itemize{
\item \code{\link{oneMeanTTest}} - Perform one-sample t-test with diagnostics
\item \code{\link{print.oneMeanTTest}} - Pretty print results
}
}

\section{Key Features}{

\itemize{
\item Wallace approximation for critical values
\item 4-term Hastings approximation for p-values
\item Shapiro-Wilk normality test
\item 1.5 * IQR outlier detection
\item Diagnostic plots (Q-Q plot, histogram, boxplot)
\item Detailed interpretation of results
}
}

\section{Getting Started}{

Load the package and run a basic test:

\code{library(OneMeanDream)}

\code{# Sample data}
\code{set.seed(123)}
\code{data <- data.frame(x = rnorm(30, mean = 10, sd = 2))}

\code{# Run test}
\code{result <- oneMeanTTest("x", mu = 10, ha = "not equal", alpha = 0.05, data = data)}

\code{# View results}
\code{print(result)}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://kingssibal.github.io/OneMeanDream/}
}

}
\author{
\strong{Maintainer}: Reijel Agub \email{rcagub@up.edu.ph}

Authors:
\itemize{
  \item Arianna Tagulinao
  \item Cindy Barcena
  \item Francesca Anne Baclig
  \item Josiah Tan
  \item Junel Rey
  \item Keano Pastrana
  \item Matt Andrei Miclat
  \item Mica Gonzales
  \item Zan Eros Policarpio
}

}
\keyword{internal}
