% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneMeanTTest.R
\name{print.oneMeanTTest}
\alias{print.oneMeanTTest}
\title{Print Method for oneMeanTTest Objects}
\usage{
\method{print}{oneMeanTTest}(x, ...)
}
\arguments{
\item{x}{An object of class "oneMeanTTest".}

\item{...}{Additional arguments (ignored).}
}
\description{
Pretty printing for oneMeanTTest results with assumption checks and interpretation.
}
